fix: Change sleeper league id from integer to bigint to hold lengthy sleeper league id integers

+++ db/migrate/20241005001853_create_user_sleeper_league.rb
@@ -2,7 +2,7 @@ class CreateUserSleeperLeague < ActiveRecord::Migration[7.1]
   def change
     create_table :user_sleeper_leagues do |t|
       t.string :sleeper_league_name
-      t.integer :sleeper_league_id
+      t.bigint :sleeper_league_id
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
